requirements.txt
langchain
langchain-community
langchain-huggingface
openai
faiss-cpu
sentence-transformers
huggingface-hub
transformers
fastapi
uvicorn[standard]
pydantic>=2.0
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
langchain_postgres
psycopg[binary,async]
pypdf
python-docx
duckduckgo-search
python-multipart